apps/agent_worker/requirements.txt
# LiveKit Agents + OpenAI Realtime plugin
livekit-agents[openai]~=1.3
livekit-plugins-noise-cancellation~=0.2
python-dotenv
requests
